moved ComponentFactories to another folder and got rid of reflections
